fix app detail field type convert and other field change and app xml deserialize test

+++ FrameMobile/tool/AppStores/BaiduAppStoreCap/Invocation/Content.cs
@@ -8,7 +8,7 @@ namespace BaiduAppStoreCap
 {
     public class Content : InvocationBase
     {
-        public int AppId { get; set; }
+        public long AppId { get; set; }
 
         public string PackageName { get; set; }
